pyproject.toml
[tool.poetry]
name = "safeack-backend"
version = "0.1.0"
description = ""
authors = ["Dhrumil Mistry <56185972+dmdhrumilmistry@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pyjwt = "^2.8.0"
python-decouple = "^3.8"
pydantic = {extras = ["email"], version = "^2.5.3"}

[tool.poetry.group.dev.dependencies]
poetry = "^1.7.1"
httpx = "^0.26.0"
pytest = "^7.4.4"

[tool.poetry.scripts]
safeack-backend = "safeack_backend.__main__:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
